+++ experiments/CRP2_CPE/setup.py
@@ -3,48 +3,8 @@ from typing import Dict
 import numpy as np
 import pandas as pd
 
-import src.utils.petab as pet
-import src.models.cpe as cpe
-
-##########################
-### DEFINED PETAB DATA ###
-##########################
-### All functions should take no arguments and return a pet.PetabData object
-
-
-def exp_0() -> pet.PetabData:
-
-    t_eval = list(np.arange(0, 1, 0.1, dtype=float))
-    fA0s = np.array([0.25, 0.5, 0.75, 0.1], dtype=float)
-    cM0s = np.array([1.0, 1.0, 1.0, 1.0], dtype=float)
-
-    cond_df = create_conditions_df(fA0s, cM0s)
-    obs_df = default_observables()
-    param_df = default_parameters()
-    empty_meas_df = pet.define_empty_measurements(obs_df, cond_df, t_eval)
-
-    return pet.PetabData(
-        obs_df=obs_df, cond_df=cond_df, param_df=param_df, meas_df=empty_meas_df
-    )
-    
-def exp_1() -> pet.PetabData:
-    
-    t_eval = list(np.arange(0, 1, 0.1, dtype=float))
-    fA0s = np.array([0.25, 0.5, 0.75, 0.1], dtype=float)
-    cM0s = np.array([1.0, 1.0, 1.0, 1.0], dtype=float)
-
-    cond_df = create_conditions_df(fA0s, cM0s)
-    obs_df = default_observables()
-    fit_params = default_fit_params()
-    fit_params["KAA"].estimate = True
-    fit_params["KBA"].estimate = True
-    param_df = pet.define_parameters(fit_params)
-    
-    empty_meas_df = pet.define_empty_measurements(obs_df, cond_df, t_eval)
-
-    return pet.PetabData(
-        obs_df=obs_df, cond_df=cond_df, param_df=param_df, meas_df=empty_meas_df
-    )
+import polypesto.core.petab as pet
+import polypesto.models.cpe as cpe
 
 
 ##############################
