+++ experiments/equilibrium/setup.py
@@ -2,35 +2,8 @@ from typing import Dict
 
 import numpy as np
 import pandas as pd
-import petab.v1.C as C
 
-from src.utils import petab as pet
-
-##########################
-### DEFINED PETAB DATA ###
-##########################
-### All functions should take no arguments and return a pet.PetabData object
-
-
-def exp_0() -> pet.PetabData:
-
-    t_eval = list(np.arange(0, 5, 0.5, dtype=float))
-    A0s = np.array([1.0, 0.2, 0.5, 0.8], dtype=float)
-    B0s = np.array([0.0, 0.8, 0.5, 0.2], dtype=float)
-
-    cond_df = create_conditions_df(A0s, B0s)
-    obs_df = default_observables()
-    param_df = default_parameters()
-    empty_meas_df = pet.define_empty_measurements(obs_df, cond_df, t_eval)
-
-    return pet.PetabData(
-        obs_df=obs_df, cond_df=cond_df, param_df=param_df, meas_df=empty_meas_df
-    )
-
-
-def exp_1() -> pet.PetabData:
-    pass
-
+from polypesto.core import petab as pet
 
 ##############################
 ### PETAB HELPER FUNCTIONS ###
